Separated the HomeScreen and Home widgets for #seprationofconcerns
